hot16_statistics/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
djangorestframework==3.11.0
fancycompleter==0.9.1
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
oauthlib==3.1.0
pdbpp==0.10.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
psycopg2-binary==2.8.5
Pygments==2.6.1
pyrepl==0.9.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
wmctrl==0.3
